+++ Sources/UI/CreateGroupCell.swift
@@ -1,14 +1,24 @@
 import UIKit
 
 class CreateGroupCell: UICollectionViewCell {
-    private var image: UIImageView = {
-        let view = UIImageView()
-        view.backgroundColor = UIColor.brandColor.withAlphaComponent(0.5)
-        view.tintColor = .brandColor
-        view.clipsToBounds = true
-        view.layer.masksToBounds = true
-        view.contentMode = .scaleAspectFill
+    private var image: UIView = {
+        let view = UIView()
         view.translatesAutoresizingMaskIntoConstraints = false
+        view.backgroundColor = UIColor.brandColor.withAlphaComponent(0.5)
+        view.clipsToBounds = true
+        view.layer.cornerRadius = 56 / 2
+        view.layer.masksToBounds = true
+
+        let image = UIImageView(image: UIImage(systemName: "plus", withConfiguration: UIImage.SymbolConfiguration(pointSize: 40, weight: .semibold)))
+        image.translatesAutoresizingMaskIntoConstraints = false
+        image.tintColor = .brandColor
+        view.addSubview(image)
+
+        NSLayoutConstraint.activate([
+            image.centerXAnchor.constraint(equalTo: view.centerXAnchor),
+            image.centerYAnchor.constraint(equalTo: view.centerYAnchor),
+        ])
+
         return view
     }()
 
@@ -39,14 +49,15 @@ class CreateGroupCell: UICollectionViewCell {
         ])
 
         NSLayoutConstraint.activate([
-            image.heightAnchor.constraint(equalToConstant: frame.size.width),
-            image.widthAnchor.constraint(equalToConstant: frame.size.width),
-            image.topAnchor.constraint(equalTo: contentView.topAnchor),
+            image.topAnchor.constraint(equalTo: contentView.topAnchor, constant: 4),
+            image.leftAnchor.constraint(equalTo: contentView.leftAnchor, constant: 4),
+            image.rightAnchor.constraint(equalTo: contentView.rightAnchor, constant: -4),
+            image.heightAnchor.constraint(equalTo: image.widthAnchor),
             image.centerXAnchor.constraint(equalTo: contentView.centerXAnchor),
         ])
 
         NSLayoutConstraint.activate([
-            title.topAnchor.constraint(equalTo: image.bottomAnchor, constant: 4),
+            title.topAnchor.constraint(equalTo: image.bottomAnchor, constant: 8),
             title.leftAnchor.constraint(equalTo: contentView.leftAnchor),
             title.rightAnchor.constraint(equalTo: contentView.rightAnchor),
         ])
@@ -54,12 +65,5 @@ class CreateGroupCell: UICollectionViewCell {
 
     required init?(coder _: NSCoder) {
         fatalError("init(coder:) has not been implemented")
-    }
-
-    override func layoutSubviews() {
-        super.layoutSubviews()
-
-        image.image = UIImage(systemName: "plus", withConfiguration: UIImage.SymbolConfiguration(weight: .semibold))
-        image.layer.cornerRadius = frame.width / 2
     }
 }